*changed HashMap to THashMap (trove-api) *introduced new function isEmpty in Attribute.Entity (isEmpty() should be used instead of getValue()==null -> this avoids export of THashMap to all models)
git-svn-id: http://svn.geogr.uni-jena.de/svn/jams/trunk@2762 cee23a32-52fb-0310-bcb1-fe94ad5b3aa2

+++ JAMSmain/src/jams/data/JAMSDouble.java
@@ -39,7 +39,7 @@ public class JAMSDouble implements Attribute.Double {
     }
 
     JAMSDouble(double value) {
-        this.value = value;
+        this.value = (double)value;
     }
 
     /*
@@ -60,11 +60,11 @@ public class JAMSDouble implements Attribute.Double {
     }
 
     public void setValue(double value) {
-        this.value = value;
+        this.value = (double)value;
     }
 
     public void setValue(String value) {
-        this.value = Double.parseDouble(value);
+        this.value = (double)Double.parseDouble(value);
     }
 
     public boolean equals(Object other) {
